running all test against a mocked response
the server response was recorded and can be played back for the unit test. Still, changing code or tests would need manually tweaking of the VCR setup code, hmm. not so nice.

Also, all the randomness from the tests was replace with deterministic values to be able to use the pre-recorded server responses. Also not so nice, but better than depending on some external component?

I will keep this on an extra branch and not on master for now.