Added validation of the path when destroying.
From time to time it can happen that destroy(!) gets called with nil or empty string. This leads to deletion of '/' which normally is not as intended. We now demand that '/' has to be explicitely stated if the root node should be deleted.

destroy! nil
or
destroy ''

normally is a programming error with potentially catastrophic consequences.